[refactor]영양제 상세 정보 entity 변경

+++ back/yourpilling/src/main/java/com/ssafy/yourpilling/pill_heeju/model/service/vo/NutritionVo.java
@@ -8,7 +8,7 @@ import lombok.Value;
 public class NutritionVo {
     Long nutritionId;
     String nutrition;
-    float amount;
+    Double amount;
     String unit;
-    float includePercent;
+    Double includePercent;
 }
